requirements.txt
elasticsearch==7.16.2
requests==2.28.1
lxml==4.9.1
bs4==0.0.1
python-dotenv==0.20.0
pandas==1.4.3
numpy==1.23.0


